Add the schema validation to the manifest loader
