+++ biothings/hub/dataplugin/loaders/dumper_meta.py
@@ -2,6 +2,7 @@ import os
 from urllib3.exceptions import InsecureRequestWarning
 from urllib3 import disable_warnings
 import biothings, config
+
 biothings.config_for_app(config)
 from config import DATA_ARCHIVE_ROOT
 
@@ -12,11 +13,13 @@ disable_warnings(InsecureRequestWarning)
 
 import biothings.hub.dataload.dumper
 
+
 class ManifestDumper(type):
     """
     Metaclass for defining the class structure
     for a manifest based dumper
     """
+
     SRC_NAME = None
     DATA_PLUGIN_FOLDER = None
     SRC_ROOT_FOLDER = None
@@ -27,6 +30,7 @@ class ManifestDumper(type):
 
     def __new__(cls, name: str, bases: list[str], attrs: list[str]):
         return super().__new__(cls, name, bases, attrs)
+
 
 class AssistedDumper(metaclass=ManifestDumper):
     pass
